+++ internal/split/iter_test.go
@@ -1,4 +1,4 @@
-package lines
+package split
 
 import (
 	"errors"
